Created Session Checker for Servlets

+++ src/com/servlets/DeleteBook.java
@@ -2,17 +2,53 @@ package com.servlets;
 
 
 import java.io.IOException;
+import java.io.PrintWriter;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.dao.BookDao;
 @WebServlet("/DeleteBook")
 public class DeleteBook extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		BookDao.delete(request.getParameter("callno"));
-		response.sendRedirect("ViewBook");
+		
+		response.setContentType("text/html");
+		PrintWriter out=response.getWriter();
+		
+		out.print("<!DOCTYPE html>");
+		out.print("<html>");
+		out.println("<head>");
+		out.println("<title>Librarian Section</title>");
+		out.println("<link rel='stylesheet' href='bootstrap.min.css'/>");
+		out.println("<link rel='stylesheet' href='index.css'/>");
+		out.println("</head>");
+		
+		HttpSession session = request.getSession();
+		
+		if (session.getAttribute("librarianemail") != null) {
+			
+			BookDao.delete(request.getParameter("callno"));
+			response.sendRedirect("ViewBook");
+		
+		}
+			
+		else {
+		
+		request.getRequestDispatcher("navhome.html").include(request, response);
+		out.println("<div class='container'>");
+		request.getRequestDispatcher("librarianloginform.html").include(request, response);
+		out.println("</div>");
+		
+		request.getRequestDispatcher("footer.html").include(request, response);
+		
+		
+		}		
+		
+		out.close();
+
 	}
 }
