+++ src/com/servlets/ViewBook.java
@@ -1,6 +1,4 @@
 package com.servlets;
-
-
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.List;
@@ -10,11 +8,14 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
+import com.authfunctions.*;
 import com.beans.BookBean;
 import com.beans.LibrarianBean;
 import com.dao.BookDao;
 import com.dao.LibrarianDao;
+
 @WebServlet("/ViewBook")
 public class ViewBook extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -28,21 +29,32 @@ public class ViewBook extends HttpServlet {
 		out.println("<link rel='stylesheet' href='bootstrap.min.css'/>");
 		out.println("</head>");
 		out.println("<body>");
-		request.getRequestDispatcher("navlibrarian.html").include(request, response);
 		
-		out.println("<div class='container'>");
+		HttpSession session = request.getSession();
 		
-		List<BookBean> list=BookDao.view();
+		if (session.getAttribute("librarianemail") != null)
+		{
 		
-		out.println("<table class='table table-bordered table-striped'>");
-		out.println("<tr><th>Callno</th><th>Name</th><th>Author</th><th>Publisher</th><th>Quantity</th><th>Issued</th><th>Delete</th></tr>");
-		for(BookBean bean:list){
-			out.println("<tr><td>"+bean.getIsbn()+"</td><td>"+bean.getTitle()+"</td><td>"+bean.getAuthor()+"</td><td>"+bean.getPublisher()+"</td><td>"+bean.getQuantity()+"</td><td>"+bean.getIssued()+"</td><td><a href='DeleteBook?callno="+bean.getIsbn()+"'>Delete</a></td></tr>");
+			request.getRequestDispatcher("navlibrarian.html").include(request, response);
+			
+			out.println("<div class='container'>");
+			
+			List<BookBean> list=BookDao.view();
+			
+			out.println("<table class='table table-bordered table-striped'>");
+			out.println("<tr><th>Callno</th><th>Name</th><th>Author</th><th>Publisher</th><th>Quantity</th><th>Issued</th><th>Delete</th></tr>");
+			for(BookBean bean:list){
+				out.println("<tr><td>"+bean.getIsbn()+"</td><td>"+bean.getTitle()+"</td><td>"+bean.getAuthor()+"</td><td>"+bean.getPublisher()+"</td><td>"+bean.getQuantity()+"</td><td>"+bean.getIssued()+"</td><td><a href='DeleteBook?callno="+bean.getIsbn()+"'>Delete</a></td></tr>");
+			}
+			out.println("</table>");
+			
+			out.println("</div>");
 		}
-		out.println("</table>");
 		
-		out.println("</div>");
-		
+		else
+		{
+			new com.authfunctions.LibraryLogin(request,response,out);
+		}
 		
 		request.getRequestDispatcher("footer.html").include(request, response);
 		out.close();
